Cut, copy, and paste menu entries
